refacto controllers, add admin security on admin routes and implement new forms and queries for different actions
