+++ app/Resources/views/default/index.html.twig
@@ -9,11 +9,11 @@
     <div class="row">
         <a href="{{ path('task_create') }}" class="btn btn-success">Créer une nouvelle tâche</a>
         <a href="{{ path('task_list') }}" class="btn btn-info">Consulter la liste des tâches à faire</a>
-        <a href="" class="btn btn-secondary">Consulter la liste des tâches terminées</a>
+        <a href="{{ path('task_list_done') }}" class="btn btn-primary">Consulter la liste des tâches terminées</a>
     </div>
     {% if app.user and 'ROLE_ADMIN' in app.user.roles %}
         <div class="row">
-            <a href="{{ path('task_list_orphans') }}" class="btn btn-primary">Consulter les tâches orphelines</a>
+            <a href="{{ path('task_list_anonymous') }}" class="btn btn-primary">Consulter les tâches anonymes</a>
             <a href="{{ path('user_list') }}" class="btn btn-primary">Consulter les utilisateurs</a>
         </div>   
     {% endif %}
