+++ app/Resources/views/security/signup.html.twig
@@ -1,0 +1,14 @@
+{% extends 'base.html.twig' %}
+
+{% block header_title %}<h1>Inscription</h1>{% endblock %}
+{% block header_img %}{% endblock %}
+
+{% block body %}
+    <div class="row">
+        {{ form_start(form, {'action' : path('signup')}) }}
+            {{ form_widget(form) }}
+
+            <button type="submit" class="btn btn-success pull-right">S'inscrire</button>
+        {{ form_end(form) }}
+    </div>
+{% endblock %}
