+++ app/Resources/views/task/list.html.twig
@@ -30,8 +30,6 @@
         {% else %}
             <div class="alert alert-warning" role="alert">
                 Il n'y a pas encore de tâche enregistrée.
-                {# Code redondant #}
-                {# <a href="{{ path('task_create') }}" class="btn btn-warning pull-right">Créer une tâche</a> #}
             </div>
         {% endfor %}
     </div>
