+++ src/AppBundle/Controller/Authentication/LoginController.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace AppBundle\Controller;
+namespace AppBundle\Controller\Authentication;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Symfony\Component\HttpFoundation\Request;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Symfony\Component\HttpFoundation\Request;
 
-class SecurityController extends Controller
+class LoginController extends Controller
 {
     /**
      * @Route("/login", name="login")
