+++ src/AppBundle/Controller/Task/DeleteTaskController.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace AppBundle\Controller\Task;
+
+use AppBundle\Entity\Task;
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+
+class DeleteTaskController extends Controller
+{
+    /**
+     * @Route("/tasks/{id}/delete", name="task_delete")
+     */
+    public function deleteTaskAction(Task $task)
+    {
+        $em = $this->getDoctrine()->getManager();
+        $em->remove($task);
+        $em->flush();
+
+        $this->addFlash('success', 'La tâche a bien été supprimée.');
+
+        return $this->redirectToRoute('task_list');
+    }
+}
