+++ src/AppBundle/Controller/Task/EditTaskController.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace AppBundle\Controller\Task;
+
+use AppBundle\Entity\Task;
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\HttpFoundation\Request;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+
+class EditTaskController extends Controller
+{
+    /**
+     * @Route("/tasks/{id}/edit", name="task_edit")
+     */
+    public function editAction(Task $task, Request $request)
+    {
+        $form = $this->createForm(TaskType::class, $task);
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $this->getDoctrine()->getManager()->flush();
+
+            $this->addFlash('success', 'La tâche a bien été modifiée.');
+
+            return $this->redirectToRoute('task_list');
+        }
+
+        return $this->render('task/edit.html.twig', [
+            'form' => $form->createView(),
+            'task' => $task,
+        ]);
+    }
+}
