+++ src/AppBundle/Controller/Task/ToggleTaskController.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace AppBundle\Controller\Task;
+
+use AppBundle\Entity\Task;
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+
+class ToggleTaskController extends Controller
+{
+    /**
+     * @Route("/tasks/{id}/toggle", name="task_toggle")
+     */
+    public function toggleTaskAction(Task $task)
+    {
+        $task->toggle(!$task->isDone());
+        $this->getDoctrine()->getManager()->flush();
+
+        $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+
+        return $this->redirectToRoute('task_list');
+    }
+}
