+++ src/AppBundle/Controller/User/CreateUserController.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace AppBundle\Controller\User;
+
+use AppBundle\Entity\User;
+use AppBundle\Form\UserType;
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Symfony\Component\BrowserKit\Request;
+
+class CreateUserController extends Controller
+{
+    /**
+     * @Route("/admin/users/create", name="user_create")
+     */
+    public function createAction(Request $request)
+    {
+        $user = new User();
+
+        $form = $this->createForm(UserType::class, $user);
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $em = $this->getDoctrine()->getManager();
+            $password = $this->get('security.password_encoder')->encodePassword($user, $user->getPassword());
+            $user->setPassword($password);
+            $user->setRoles(['ROLE_USER']);
+
+            $em->persist($user);
+            $em->flush();
+
+            $this->addFlash('success', "L'utilisateur a bien été ajouté.");
+
+            // changed the redirection to avoid user on user list page
+            return $this->redirectToRoute('user_list');
+        }
+
+        return $this->render('user/create.html.twig', ['form' => $form->createView()]);
+    }
+}
