+++ src/AppBundle/Controller/User/EditUserController.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace AppBundle\Controller\User;
+
+use AppBundle\Entity\User;
+use AppBundle\Form\UserType;
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Symfony\Component\HttpFoundation\Request;
+
+class EditUserController extends Controller
+{
+    
+    /**
+     * @Route("/admin/users/{id}/edit", name="user_edit")
+     */
+    public function editAction(User $user, Request $request)
+    {
+        $form = $this->createForm(UserType::class, $user);
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $password = $this->get('security.password_encoder')->encodePassword($user, $user->getPassword());
+            $user->setPassword($password);
+
+            $this->getDoctrine()->getManager()->flush();
+
+            $this->addFlash('success', "L'utilisateur a bien été modifié");
+
+            return $this->redirectToRoute('user_list');
+        }
+
+        return $this->render('user/edit.html.twig', ['form' => $form->createView(), 'user' => $user]);
+    }
+}
