+++ src/AppBundle/Controller/User/ListUserController.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace AppBundle\Controller\User;
+
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+
+class ListUserController extends Controller
+{
+    /**
+     * @Route("/admin/users", name="user_list")
+     */
+    public function listAction()
+    {
+        return $this->render('user/list.html.twig', ['users' => $this->getDoctrine()->getRepository('AppBundle:User')->findAll()]);
+    }
+
+}
