--- src/AppBundle/Controller/UserController.php
@@ -1,89 +1,0 @@
-<?php
-
-namespace AppBundle\Controller;
-
-use AppBundle\Entity\User;
-use AppBundle\Form\SigninType;
-use AppBundle\Form\UserType;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Symfony\Component\HttpFoundation\Request;
-
-class UserController extends Controller
-{
-    /**
-     * @Route("/users", name="user_list")
-     */
-    public function listAction()
-    {
-        return $this->render('user/list.html.twig', ['users' => $this->getDoctrine()->getRepository('AppBundle:User')->findAll()]);
-    }
-
-    /**
-     * @Route("/users/create", name="user_create")
-     */
-    public function createAction(Request $request)
-    {
-        $user = new User();
-
-        $form = $this->createForm(SigninType::class, $user);
-
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $password = $this->get('security.password_encoder')->encodePassword($user, $user->getPassword());
-            $user->setPassword($password);
-            $user->setRoles(['ROLE_USER']);
-
-            $em->persist($user);
-            $em->flush();
-
-            $this->addFlash('success', "L'utilisateur a bien été ajouté.");
-
-            // changed the redirection to avoid user on user list page
-            return $this->redirectToRoute('homepage');
-        }
-
-        return $this->render('user/create.html.twig', ['form' => $form->createView()]);
-    }
-
-    /**
-     * @Route("/users/{id}/edit", name="user_edit")
-     */
-    public function editAction(User $user, Request $request)
-    {
-        $form = $this->createForm(UserType::class, $user);
-
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $password = $this->get('security.password_encoder')->encodePassword($user, $user->getPassword());
-            $user->setPassword($password);
-
-            $this->getDoctrine()->getManager()->flush();
-
-            $this->addFlash('success', "L'utilisateur a bien été modifié");
-
-            return $this->redirectToRoute('user_list');
-        }
-
-        return $this->render('user/edit.html.twig', ['form' => $form->createView(), 'user' => $user]);
-    }
-
-    /**
-     * @Route("/users/{id}/delete", name="user_delete")
-     */
-    public function deleteAction($id)
-    {
-        $user = $this->getDoctrine()->getRepository('AppBundle:User')->find((int) $id);
-
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($user);
-        $em->flush();
-
-        $this->addFlash('success', 'L\' utilisateur à bien été supprimer');
-
-        return $this->redirectToRoute('user_list');
-    }
-}
