+++ src/AppBundle/Form/UserType.php
@@ -16,13 +16,13 @@ class UserType extends AbstractType
     {
         $builder
             ->add('username', TextType::class, ['label' => "Nom d'utilisateur"])
-            // ->add('password', RepeatedType::class, [
-            //     'type' => PasswordType::class,
-            //     'invalid_message' => 'Les deux mots de passe doivent correspondre.',
-            //     'required' => true,
-            //     'first_options'  => ['label' => 'Mot de passe'],
-            //     'second_options' => ['label' => 'Tapez le mot de passe à nouveau'],
-            // ])
+            ->add('password', RepeatedType::class, [
+                'type' => PasswordType::class,
+                'invalid_message' => 'Les deux mots de passe doivent correspondre.',
+                'required' => true,
+                'first_options'  => ['label' => 'Mot de passe'],
+                'second_options' => ['label' => 'Tapez le mot de passe à nouveau'],
+            ])
             ->add('email', EmailType::class, ['label' => 'Adresse email'])
             ->add('roles', ChoiceType::class, [
                 'label' => 'Roles de l\'utilisateur',
